+++ src/tests/collections/equation/color/animations.test.js
@@ -1,0 +1,5 @@
+import { staticTester } from '../../../staticTester';
+
+const { getShapes, updates } = require('./animations');
+
+staticTester('Collection: Equation Color - Animations', getShapes, updates);