+++ src/tests/collections/equation/color/static.test.js
@@ -1,0 +1,7 @@
+import { staticTester } from '../../../staticTester';
+
+const {
+  getShapes, updates, getValues, move,
+} = require('./static');
+
+staticTester('Collection: Equation - Lines', getShapes, updates, getValues, move);